+++ bot.py
@@ -35,11 +35,6 @@ import traceback
 
 sys.path.append("ad_discordbot")
 
-from database import Database
-
-#################################################################
-#################### DISCORD / BOT STARTUP ######################
-#################################################################
 # Start logger
 logging.basicConfig(format='%(levelname)s [%(asctime)s]: %(message)s (Line: %(lineno)d in %(funcName)s, %(filename)s )',
                     datefmt='%Y-%m-%d %H:%M:%S', 
@@ -52,6 +47,17 @@ handler = logging.handlers.RotatingFileHandler(
     maxBytes=32 * 1024 * 1024,  # 32 MiB
     backupCount=5,  # Rotate through 5 files
 )
+
+from database import Database
+from utils_shared import task_semaphore
+from utils_misc import fix_dict, update_dict, sum_update_dict, update_dict_matched_keys
+from utils_discord import ireply, send_long_message
+from utils_files import load_file, merge_base, save_yaml_file
+
+#################################################################
+#################### DISCORD / BOT STARTUP ######################
+#################################################################
+
 
 # Resolve legacy config method
 class Config:
@@ -489,179 +495,6 @@ async def process_tasks_in_background():
     while True:
         task = await bg_task_queue.get()
         await task
-
-#################################################################
-######################## MISC FUNCTIONS #########################
-#################################################################
-# Function to load .json, .yml or .yaml files
-def load_file(file_path):
-    try:
-        file_suffix = Path(file_path).suffix.lower()
-
-        if file_suffix in [".json", ".yml", ".yaml"]:
-            with open(file_path, 'r', encoding='utf-8') as file:
-                if file_suffix in [".json"]:
-                    data = json.load(file)
-                else:
-                    data = yaml.safe_load(file)
-            return data
-        else:
-            logging.error(f"Unsupported file format: {file_suffix}")
-            return None
-    except FileNotFoundError:
-        logging.error(f"File not found: {file_suffix}")
-        return None
-    except Exception as e:
-        logging.error(f"An error occurred while reading {file_path}: {str(e)}")
-        return None
-
-def merge_base(newsettings, basekey):
-    def deep_update(original, update):
-        for key, value in update.items():
-            if isinstance(value, dict) and key in original and isinstance(original[key], dict):
-                deep_update(original[key], value)
-            else:
-                original[key] = value
-    try:
-        base_settings = load_file('ad_discordbot/dict_base_settings.yaml')
-        keys = basekey.split(',')
-        current_dict = base_settings
-        for key in keys:
-            if key in current_dict:
-                current_dict = current_dict[key].copy()
-            else:
-                return None
-        deep_update(current_dict, newsettings) # Recursively update the dictionary
-        return current_dict
-    except Exception as e:
-        logging.error(f"Error loading ad_discordbot/dict_base_settings.yaml ({basekey}): {e}")
-        return newsettings
-
-# Send message response to user's interaction command
-async def ireply(i, process):
-    try:
-        if task_semaphore.locked(): # If a queued item is currently being processed
-            ireply = await i.reply(f'Your {process} request was added to the task queue', ephemeral=True, delete_after=5)
-            # del_time = 5
-        else:
-            ireply = await i.reply(f'Processing your {process} request', ephemeral=True, delete_after=3)
-        #     del_time = 1
-    except Exception as e:
-        logging.error(f"Error sending message response to user's interaction command: {e}")
-
-# Adds missing keys/values
-def fix_dict(set, req):
-    for k, req_v in req.items():
-        if k not in set:
-            set[k] = req_v
-        elif isinstance(req_v, dict):
-            fix_dict(set[k], req_v)
-    return set
-
-# Updates matched keys, AND adds missing keys
-def update_dict(d, u):
-    for k, v in u.items():
-        if isinstance(v, dict):
-            d[k] = update_dict(d.get(k, {}), v)
-        else:
-            d[k] = v
-    for k in d.keys() - u.keys():
-        u[k] = d[k]
-    return u
-
-# Updates matched keys, AND adds missing keys, BUT sums together number values
-def sum_update_dict(d, u):
-    def get_decimal_places(value):
-        # Function to get the number of decimal places in a float.
-        if isinstance(value, float):
-            return len(str(value).split('.')[1])
-        else:
-            return 0
-    for k, v in u.items():
-        if isinstance(v, dict):
-            d[k] = sum_update_dict(d.get(k, {}), v)
-        elif isinstance(v, (int, float)) and not isinstance(v, bool):
-            current_value = d.get(k, 0)
-            max_decimal_places = max(get_decimal_places(current_value), get_decimal_places(v))
-            d[k] = round(current_value + v, max_decimal_places)
-        else:
-            d[k] = v
-    for k in d.keys() - u.keys():
-        u[k] = d[k]
-    return u
-
-# Updates matched keys, but DOES NOT add missing keys
-def update_dict_matched_keys(d, u):
-    for k, v in u.items():
-        if isinstance(v, dict):
-            d[k] = update_dict(d.get(k, {}), v)
-        else:
-            d[k] = v
-    return d
-
-def save_yaml_file(file_path, data):
-    try:
-        with open(file_path, 'w') as file:
-            yaml.dump(data, file, encoding='utf-8', default_flow_style=False, width=float("inf"), sort_keys=False)
-    except Exception as e:
-        logging.error(f"An error occurred while saving {file_path}: {str(e)}")
-
-async def send_long_message(channel, message_text):
-    """ Splits a longer message into parts while preserving sentence boundaries and code blocks """
-    activelang = ''
-
-    # Helper function to ensure even pairs of code block markdown
-    def ensure_even_code_blocks(chunk_text, code_block_inserted):
-        nonlocal activelang  # Declare activelang as nonlocal to modify the global variable
-        code_block_languages = ["asciidoc", "autohotkey", "bash", "coffeescript", "cpp", "cs", "css", "diff", "fix", "glsl", "ini", "json", "md", "ml", "prolog", "ps", "py", "tex", "xl", "xml", "yaml", "html"]
-        code_block_count = chunk_text.count("```")
-        if code_block_inserted:
-            # If a code block was inserted in the previous chunk, add a leading set of "```"
-            chunk_text = f"```{activelang}\n" + chunk_text
-            code_block_inserted = False  # Reset the code_block_inserted flag
-        code_block_count = chunk_text.count("```")
-        if code_block_count % 2 == 1:
-            # Check last code block for syntax like "```yaml"
-            last_code_block_index = chunk_text.rfind("```")
-            last_code_block = chunk_text[last_code_block_index + len("```"):].strip()
-            for lang in code_block_languages:
-                if (last_code_block.lower()).startswith(lang):
-                    activelang = lang
-                    break  # Stop checking if a match is found
-            # If there is an odd number of code blocks, add a closing set of "```"
-            chunk_text += "```"
-            code_block_inserted = True
-        return chunk_text, code_block_inserted
-
-    if len(message_text) <= 1980:
-        sent_message = await channel.send(message_text)
-    else:
-        code_block_inserted = False  # Initialize code_block_inserted to False
-        while message_text:
-            # Find the last occurrence of either a line break or the end of a sentence
-            last_line_break = message_text.rfind("\n", 0, 1980)
-            last_sentence_end = message_text.rfind(". ", 0, 1980)
-            # Determine the index to split the string
-            if last_line_break >= 0 and last_sentence_end >= 0:
-                # If both a line break and a sentence end were found, choose the one that occurred last
-                chunk_length = max(last_line_break, last_sentence_end) + 1
-            elif last_line_break >= 0:
-                # If only a line break was found, use it as the split point
-                chunk_length = last_line_break + 1
-            elif last_sentence_end >= 0:
-                # If only a sentence end was found, use it as the split point
-                chunk_length = last_sentence_end + 2  # Include the period and space
-            else:
-                chunk_length = 1980 # If neither was found, split at the maximum limit of 2000 characters
-            chunk_text = message_text[:chunk_length]
-            chunk_text, code_block_inserted = ensure_even_code_blocks(chunk_text, code_block_inserted)
-            sent_message = await channel.send(chunk_text)
-            message_text = message_text[chunk_length:]
-            if len(message_text) <= 1980:
-                # Send the remaining text as a single chunk if it's shorter than or equal to 2000 characters
-                chunk_text, code_block_inserted = ensure_even_code_blocks(message_text, code_block_inserted)
-                sent_message = await channel.send(chunk_text)
-                break
 
 #################################################################
 ########################## BOT STARTUP ##########################
@@ -2278,8 +2111,6 @@ def update_mention(user_id, last_resp=''):
 
 flow_event = asyncio.Event()
 flow_queue = asyncio.Queue()
-
-task_semaphore = asyncio.Semaphore(1)
 
 #################################################################
 ########################## QUEUED FLOW ##########################
