+++ utils_discord.py
@@ -1,0 +1,107 @@
+from datetime import datetime, timedelta
+from pathlib import Path
+import asyncio
+import random
+import logging
+import logging.handlers
+import json
+import re
+import glob
+import os
+import warnings
+import discord
+from discord.ext import commands
+from discord import app_commands, File
+from discord.ext.commands.context import Context
+import typing
+import torch
+import io
+import base64
+import yaml
+from PIL import Image, PngImagePlugin
+import requests
+import sqlite3
+import pprint
+import aiohttp
+import math
+import time
+from itertools import product
+from threading import Lock, Thread
+from pydub import AudioSegment
+import copy
+from shutil import copyfile
+import sys
+import traceback
+
+from utils_shared import task_semaphore
+
+# Send message response to user's interaction command
+async def ireply(i, process):
+    try:
+        if task_semaphore.locked(): # If a queued item is currently being processed
+            ireply = await i.reply(f'Your {process} request was added to the task queue', ephemeral=True, delete_after=5)
+            # del_time = 5
+        else:
+            ireply = await i.reply(f'Processing your {process} request', ephemeral=True, delete_after=3)
+        #     del_time = 1
+    except Exception as e:
+        logging.error(f"Error sending message response to user's interaction command: {e}")
+        
+        
+        
+async def send_long_message(channel, message_text):
+    """ Splits a longer message into parts while preserving sentence boundaries and code blocks """
+    activelang = ''
+
+    # Helper function to ensure even pairs of code block markdown
+    def ensure_even_code_blocks(chunk_text, code_block_inserted):
+        nonlocal activelang  # Declare activelang as nonlocal to modify the global variable
+        code_block_languages = ["asciidoc", "autohotkey", "bash", "coffeescript", "cpp", "cs", "css", "diff", "fix", "glsl", "ini", "json", "md", "ml", "prolog", "ps", "py", "tex", "xl", "xml", "yaml", "html"]
+        code_block_count = chunk_text.count("```")
+        if code_block_inserted:
+            # If a code block was inserted in the previous chunk, add a leading set of "```"
+            chunk_text = f"```{activelang}\n" + chunk_text
+            code_block_inserted = False  # Reset the code_block_inserted flag
+        code_block_count = chunk_text.count("```")
+        if code_block_count % 2 == 1:
+            # Check last code block for syntax like "```yaml"
+            last_code_block_index = chunk_text.rfind("```")
+            last_code_block = chunk_text[last_code_block_index + len("```"):].strip()
+            for lang in code_block_languages:
+                if (last_code_block.lower()).startswith(lang):
+                    activelang = lang
+                    break  # Stop checking if a match is found
+            # If there is an odd number of code blocks, add a closing set of "```"
+            chunk_text += "```"
+            code_block_inserted = True
+        return chunk_text, code_block_inserted
+
+    if len(message_text) <= 1980:
+        sent_message = await channel.send(message_text)
+    else:
+        code_block_inserted = False  # Initialize code_block_inserted to False
+        while message_text:
+            # Find the last occurrence of either a line break or the end of a sentence
+            last_line_break = message_text.rfind("\n", 0, 1980)
+            last_sentence_end = message_text.rfind(". ", 0, 1980)
+            # Determine the index to split the string
+            if last_line_break >= 0 and last_sentence_end >= 0:
+                # If both a line break and a sentence end were found, choose the one that occurred last
+                chunk_length = max(last_line_break, last_sentence_end) + 1
+            elif last_line_break >= 0:
+                # If only a line break was found, use it as the split point
+                chunk_length = last_line_break + 1
+            elif last_sentence_end >= 0:
+                # If only a sentence end was found, use it as the split point
+                chunk_length = last_sentence_end + 2  # Include the period and space
+            else:
+                chunk_length = 1980 # If neither was found, split at the maximum limit of 2000 characters
+            chunk_text = message_text[:chunk_length]
+            chunk_text, code_block_inserted = ensure_even_code_blocks(chunk_text, code_block_inserted)
+            sent_message = await channel.send(chunk_text)
+            message_text = message_text[chunk_length:]
+            if len(message_text) <= 1980:
+                # Send the remaining text as a single chunk if it's shorter than or equal to 2000 characters
+                chunk_text, code_block_inserted = ensure_even_code_blocks(message_text, code_block_inserted)
+                sent_message = await channel.send(chunk_text)
+                break
