+++ utils_shared.py
@@ -1,0 +1,3 @@
+import asyncio
+
+task_semaphore = asyncio.Semaphore(1)
